Organize backend leaderboard lifecycles
