+++ apps/backend/src/entrypoints/lambda-community-leaderboard-snapshot.ts
@@ -7,7 +7,7 @@ import {
   normalizeCaughtError,
   wrapBackendHandler,
 } from "../observability/sentry";
-import { LEADERBOARD_SNAPSHOT_METRIC_VERSION } from "../community/leaderboard/leaderboardWindows";
+import { LEADERBOARD_SNAPSHOT_METRIC_VERSION } from "../community/leaderboard/progress/leaderboardWindows";
 import { withTransientDatabaseRetry } from "../database/transient";
 
 initializeBackendSentry("community-leaderboard-snapshot");
@@ -21,13 +21,13 @@ type CommunityLeaderboardSnapshotResponse = Readonly<{
 }>;
 
 type CommunityLeaderboardSnapshotRuntime = Readonly<{
-  generateLeaderboardSnapshots: typeof import("../community/leaderboard/leaderboardSnapshots").generateLeaderboardSnapshots;
+  generateLeaderboardSnapshots: typeof import("../community/leaderboard/progress/leaderboardSnapshots").generateLeaderboardSnapshots;
 }>;
 
 let communityLeaderboardSnapshotRuntimePromise: Promise<CommunityLeaderboardSnapshotRuntime> | null = null;
 
 async function createCommunityLeaderboardSnapshotRuntime(): Promise<CommunityLeaderboardSnapshotRuntime> {
-  const { generateLeaderboardSnapshots } = await import("../community/leaderboard/leaderboardSnapshots");
+  const { generateLeaderboardSnapshots } = await import("../community/leaderboard/progress/leaderboardSnapshots");
   return {
     generateLeaderboardSnapshots,
   };
