+++ apps/ios/Flashcards/Flashcards/Progress/Model/ProgressLeaderboardTypes.swift
@@ -4,7 +4,7 @@ private let progressLeaderboardMaximumGapRowCount: Int = 2
 
 // Wire contract for GET /me/progress/leaderboard.
 // Keep aligned with api/src/openapi.yaml and
-// apps/backend/src/community/leaderboard/progressLeaderboard.ts.
+// apps/backend/src/community/leaderboard/progress/progressLeaderboard.ts.
 enum LeaderboardWindowKey: String, Codable, CaseIterable, Identifiable, Sendable {
     case last24Hours = "last_24_hours"
     case last3Days = "last_3_days"
