+++ apps/ios/Flashcards/Flashcards/Progress/Presentation/ProgressFormatting.swift
@@ -266,7 +266,7 @@ func progressLeaderboardSectionTitle() -> String {
 }
 
 // Keep the counting rule wording aligned with the backend metric copy in
-// apps/backend/src/community/leaderboard/progressLeaderboard.ts.
+// apps/backend/src/community/leaderboard/progress/progressLeaderboard.ts.
 func progressLeaderboardInfoMessage(snapshotGeneratedAt: String?, now: Date) -> String {
     let baseMessage = String(
         localized: "progress.screen.leaderboard.info.message",
